Broke out 'MQTTPublisher' & 'GatewayCore' classes from 'gateway.py' into 'mqttpublisher.py' and 'gatewaycore.py'.
